requirements.txt
streamlit
openai
pinecone-client
matplotlib